WIP: sched/fair: filter energy_diff() based on energy_payoff value
Once the SchedTune support is enabled and the CPU bandwidth demand of a
task is boosted, we could expect increased energy consumptions which are
balanced by corresponding increases of tasks performance.
However, the current implementation of the energy_diff() function
accepts all and _only_ the schedule candidates which results into a
reduced expected system energy, which works against the boosting
strategy.

This patch links the energy_diff() function with the "energy payoff"
engine provided by SchedTune. The energy variation computed by the
energy_diff() function is now filtered using the SchedTune support to
evaluated the energy payoff for a boosted task.

With that patch, the energy_diff() function is going to reported as
"acceptable schedule candidate" only the schedule candidate which
corresponds to a positive energy_payoff.

BUG=chrome-os-partner:44828
TEST=Boot kernel on Oak.

Change-Id: Ib0a5c5f37cfd6dd4d244cac633d3a36b99bae0c6
Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Signed-off-by: Tab Fitts <tfitts1212@gmail.com>